pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "compress_comics"
version = "v0.4.13"
authors = [
  { name="TheHardew" },
]
description = "Compress images in cbz/cbr files with jpegxl, fast."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
    "Operating System :: OS Independent",
]
dependencies = [
	"patool",
	"tqdm",
]

[project.urls]
Homepage = "https://github.com/TheHardew/compress_comics"
Issues = "https://github.com/TheHardew/compress_comics/issues"

[project.scripts]
compress_comics = "compress_comics:main"

[tool.hatch.build.targets.sdist]
only-include = [
	"src/", 
	"pyproject.toml"
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
